refactor: allow accessing the google charts' chart wrapper
This is needed when composing charts.
