+++ libs/angular-google-charts/src/index.ts
@@ -2,9 +2,11 @@
  * Public API Surface of angular-google-charts
  */
 
-export * from './lib/google-chart/google-chart.component';
+export * from './lib/components/google-chart/google-chart.component';
+export * from './lib/components/chart-wrapper/chart-wrapper.component';
 export * from './lib/helpers/chart.helper';
 export * from './lib/models/events.model';
 export * from './lib/models/chart-type.model';
+export { Column } from './lib/models/chart-base.model';
 export * from './lib/script-loader/script-loader.service';
 export * from './lib/google-charts.module';
