+++ libs/angular-google-charts/src/lib/google-charts.module.ts
@@ -1,7 +1,7 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 
-import { ChartWrapperComponent } from './chart-wrapper/chart-wrapper.component';
-import { GoogleChartComponent } from './google-chart/google-chart.component';
+import { ChartWrapperComponent } from './components/chart-wrapper/chart-wrapper.component';
+import { GoogleChartComponent } from './components/google-chart/google-chart.component';
 import { GoogleChartsConfig } from './models/google-charts-config.model';
 import { GOOGLE_CHARTS_CONFIG } from './models/injection-tokens.model';
